requirements.txt
fastapi==0.115.0
matplotlib==3.8.4
mlflow==2.16.1
mlflow_skinny==2.16.1
numpy==2.1.2
pandas==2.2.3
Pillow==9.3.0
Pillow==9.0.1
Pillow==10.4.0
prefect==3.0.3
pydantic==2.9.2
pygame==2.6.1
scikit_learn==1.4.1.post1
tensorflow==2.17.0
tensorflow==2.16.1
